// apps/dashboard/src/api/generated/restaurantDashboardAPI.schemas.ts
export interface Category {
  id: string;
  /** @maxLength 100 */
  name: string;
  /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface MenuItem {
  id: string;
  categoryId: string;
  /** @maxLength 200 */
  name: string;
  description: string;
  price: string;
  available: boolean;
  popular: boolean;
  image: string | null;
  /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}